# ⚡️ Speed up function `random_block_mask_image` by 64%