@misrasaurabh1 2025-07-01
Hi @guarin, I am working with @liopeer who's been reviewing some of these optimizations offline.